Fix NoClassDefFoundError caused by refactor in the main server
